Introduce `State::Action`
Introduces `State::Action` to mange shift, goto and reduce.
`State::Shift` and `State::Reduce` are removed.
`State::Shift` is divided into `State::Action::Shift` and `State::Action::Goto`
then type declaration changes from `Shift` to `Action::Shift | Action::Goto`.
