--- lib/lrama/state/reduce.rb
@@ -1,54 +1,0 @@
-# rbs_inline: enabled
-# frozen_string_literal: true
-
-module Lrama
-  class State
-    class Reduce
-      # TODO: rbs-inline 0.10.0 doesn't support instance variables.
-      #       Move these type declarations above instance variable definitions, once it's supported.
-      #
-      # @rbs!
-      #   @item: States::Item
-      #   @look_ahead: Array[Grammar::Symbol]?
-      #   @not_selected_symbols: Array[Grammar::Symbol]
-
-      attr_reader :item #: States::Item
-      attr_reader :look_ahead #: Array[Grammar::Symbol]?
-      attr_reader :not_selected_symbols #: Array[Grammar::Symbol]
-
-      # https://www.gnu.org/software/bison/manual/html_node/Default-Reductions.html
-      attr_accessor :default_reduction #: bool
-
-      # @rbs (States::Item item) -> void
-      def initialize(item)
-        @item = item
-        @look_ahead = nil
-        @not_selected_symbols = []
-      end
-
-      # @rbs () -> Grammar::Rule
-      def rule
-        @item.rule
-      end
-
-      # @rbs (Array[Grammar::Symbol] look_ahead) -> Array[Grammar::Symbol]
-      def look_ahead=(look_ahead)
-        @look_ahead = look_ahead.freeze
-      end
-
-      # @rbs (Grammar::Symbol sym) -> Array[Grammar::Symbol]
-      def add_not_selected_symbol(sym)
-        @not_selected_symbols << sym
-      end
-
-      # @rbs () -> (::Array[Grammar::Symbol?])
-      def selected_look_ahead
-        if look_ahead
-          look_ahead - @not_selected_symbols
-        else
-          []
-        end
-      end
-    end
-  end
-end
