+++ sig/lrama/state/reduce_reduce_conflict.rbs
@@ -2,10 +2,10 @@ module Lrama
   class State
     class ReduceReduceConflict
       attr_accessor symbols: Array[Grammar::Symbol]
-      attr_accessor reduce1: State::Reduce
-      attr_accessor reduce2: State::Reduce
+      attr_accessor reduce1: State::Action::Reduce
+      attr_accessor reduce2: State::Action::Reduce
 
-      def initialize: (?symbols: Array[Grammar::Symbol], ?reduce1: State::Reduce, ?reduce2: State::Reduce) -> void
+      def initialize: (?symbols: Array[Grammar::Symbol], ?reduce1: State::Action::Reduce, ?reduce2: State::Action::Reduce) -> void
 
       def type: () -> :reduce_reduce
     end
