+++ sig/lrama/state/shift_reduce_conflict.rbs
@@ -2,10 +2,10 @@ module Lrama
   class State
     class ShiftReduceConflict
       attr_accessor symbols: Array[Grammar::Symbol]
-      attr_accessor shift: State::Shift
-      attr_accessor reduce: State::Reduce
+      attr_accessor shift: State::Action::Shift | State::Action::Goto
+      attr_accessor reduce: State::Action::Reduce
 
-      def initialize: (?symbols: Array[Grammar::Symbol], ?shift: State::Shift, ?reduce: State::Reduce) -> void
+      def initialize: (?symbols: Array[Grammar::Symbol], ?shift: State::Action::Shift, ?reduce: State::Action::Reduce) -> void
 
       def type: () -> :shift_reduce
     end
